test(service): move integration tests
