+++ oprf-test-utils/Cargo.toml
@@ -32,22 +32,30 @@ ark-ec.workspace = true
 ark-ff.workspace = true
 ark-serialize.workspace = true
 askama = { version = "0.16", optional = true }
-axum-test.workspace = true
+axum.workspace = true
+axum-test = { workspace = true, features = ["ws"] }
+ciborium.workspace = true
 clap = { workspace = true, features = ["derive", "env"], optional = true }
 eyre.workspace = true
 futures = { workspace = true }
 groth16-material = { workspace = true, optional = true }
 groth16-sol = { workspace = true, optional = true }
+http = { workspace = true }
 itertools.workspace = true
 nodes-common = { workspace = true, features = ["postgres", "test-utils"], optional = true }
+oprf-client = { package = "taceo-oprf-client", path = "../oprf-client", version = "0.10" }
 oprf-core = { package = "taceo-oprf-core", path = "../oprf-core", version = "0.6" }
 oprf-key-gen = { package = "taceo-oprf-key-gen", path = "../oprf-key-gen/", version = "0.12" }
+oprf-service = { package = "taceo-oprf-service", path = "../oprf-service/", version = "0.18" }
 oprf-types = { package = "taceo-oprf-types", path = "../oprf-types", version = "0.15", features = [
   "chain",
   "service",
 ] }
 rand.workspace = true
 rand_chacha.workspace = true
+reqwest.workspace = true
+ruint = { workspace = true, features = ["rand"] }
+serde.workspace = true
 serde_json = { workspace = true, optional = true }
 sqlx = { workspace = true, features = [
   "postgres",
@@ -63,6 +71,8 @@ tokio = { workspace = true, features = [
   "tokio-macros",
 ] }
 tokio-util = { workspace = true, optional = true }
+tungstenite = { workspace = true }
+uuid = { workspace = true, features = ["serde", "v4"] }
 
 [features]
 default = ["deploy-anvil", "postgres-test-container"]
